+++ ricaun.Revit.UI/RibbonTextBoxExtension.cs
@@ -114,7 +114,9 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TextBoxData NewTextBoxData(this RibbonPanel ribbonPanel, string targetName = null)
         {
-            if (string.IsNullOrEmpty(targetName)) targetName = nameof(TextBox);
+            if (string.IsNullOrWhiteSpace(targetName))
+                targetName = nameof(TextBox);
+
             targetName = RibbonSafeExtension.GenerateSafeButtonName(ribbonPanel, targetName, targetName);
 
             return new TextBoxData(targetName);